zmq-pub-sub--many-1/consumer/pyproject.toml
[tool.poetry]
name = "consumer"
version = "0.1.0"
description = ""
authors = ["Konstantin Freidlin <freidlinks@ya.ru>"]
readme = "README.md"
packages = [
    { include = "src" }
]

[tool.poetry.dependencies]
python = "^3.11"
pyzmq = "^26.0.3"
loguru = "^0.7.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
consume = "src.consume:main"
